+++ middleman-core/lib/middleman-core/file_renderer.rb
@@ -51,7 +51,7 @@ module Middleman
 
       # Merge per-extension options from config
       extension = File.extname(path)
-      options = opts.merge(options_for_ext(extension))
+      options = {}.merge!(opts).merge!(options_for_ext(extension))
       options[:outvar] ||= '@_out_buf'
       options[:context] = context
       options.delete(:layout)