+++ CHANGELOG.md
@@ -19,6 +19,7 @@ master
 * Fix asset_host in combination with Google Analytics snippet. #1751
 * Show an error message when git CLI is not available. #1765
 * Correctly show file names of GZIP'ed assets. #1364
+* Build file output is not parallel-ized! Use `middleman build --no-parallel` to disable.
 
 # 4.0.0